+++ lib/ui/screens/noinstance.dart
@@ -1,55 +1,68 @@
 import 'package:flutter/material.dart';
-import 'package:sct_mobile/ui/screens/onboarding_screen.dart';
+import 'package:sct_mobile/ui/shared/sizeConfig.dart';
 import 'package:sct_mobile/ui/shared/styles.dart';
 import 'package:flutter_svg/flutter_svg.dart';
-import 'package:sct_mobile/ui/shared/sizeConfig.dart';
+import 'package:url_launcher/url_launcher.dart';
 import 'package:sct_mobile/ui/screens/scanner.dart';
 import 'package:sct_mobile/ui/widgets/primarybutton.dart';
 import 'package:sct_mobile/ui/widgets/secondarybutton.dart';
+import 'package:sct_mobile/ui/widgets/tertiarybutton.dart';
 import 'package:sct_mobile/ui/screens/login.dart';
 import 'package:sct_mobile/ui/screens/menu.dart';
 import 'package:sct_mobile/ui/screens/noinstance.dart';
+import 'package:sct_mobile/ui/widgets/tertiarybutton.dart';
 
-main() {
-  runApp(
-    MaterialApp(
-      theme: ThemeData.dark(),
-      home: Home(),
-    ),
-  );
+class NoInstance extends StatefulWidget {
+  @override
+  _NoInstanceState createState() => _NoInstanceState();
 }
 
-class Home extends StatelessWidget {
+class _NoInstanceState extends State<NoInstance> {
+  _launchURL() async {
+    const url = 'https://github.com/SCT-devTeam/SCT-Backend';
+    if (await canLaunch(url)) {
+      await launch(url);
+    } else {
+      throw 'Could not launch $url';
+    }
+  }
+
   @override
   Widget build(BuildContext context) {
-    SizeConfig().init(context);
     return MaterialApp(
       title: 'Welcome to SCT',
       home: Center(
         child: Scaffold(
-          backgroundColor: Color(0xfff7c91e),
+          resizeToAvoidBottomInset: false,
+          appBar: AppBar(
+            elevation: 0,
+            backgroundColor: Color(0xff1c2841),
+            leading: IconButton(
+              icon: Icon(Icons.arrow_back_ios),
+              onPressed: () => Navigator.pop(context),
+            ),
+          ),
+          backgroundColor: Color(0xff1c2841),
           body: Center(
             child: Column(
               mainAxisAlignment: MainAxisAlignment.center,
               children: <Widget>[
-//                SizedBox(height: 10.0 * SizeConfig.blockSizeVertical),
                 SvgPicture.asset(
-                    'assets/images/sct_logo_colored_squaredWithoutBackgroundAndSubtitle.svg',
+                    'assets/images/sct_logo_colored_squaredWithoutBackgroundAndSubtitleInverted.svg',
                     height: 20.0 * SizeConfig.blockSizeVertical,
                     width: 20.0 * SizeConfig.blockSizeHorizontal),
-                SizedBox(height: 3.0 * SizeConfig.blockSizeVertical),
+                SizedBox(height: 30.0),
                 Container(
                   child: Text(
-                    "L'application à besoin de se connecter à une instance de SCT² pour fonctionner",
-                    style: sctText,
+                    "Pour utiliser l'application vous devez avoir une instance du service",
+                    style: sctSubtitleStyleMed_w,
                     textAlign: TextAlign.center,
                   ),
-                  padding:
-                      const EdgeInsets.all(3.0) * SizeConfig.blockSizeVertical,
+                  padding: const EdgeInsets.all(20.0),
                 ),
-                SizedBox(height: 50.0),
-                SecondaryButton(
-                  title: Text("Scanner le QR Code",
+                SizedBox(height: 2.0 * SizeConfig.blockSizeVertical),
+                TertiaryButton(
+                  title: Text("Commander une instance",
                       style: TextStyle(
                           fontFamily: 'Roboto',
                           fontWeight: FontWeight.w500,
@@ -58,29 +71,25 @@ class Home extends StatelessWidget {
                     context,
                     MaterialPageRoute(builder: (context) => Scanner()),
                   ),
+                  isEnabled: true,
                 ),
-                SecondaryButton(
-                  title: Text("Utiliser ses identifiants",
+                PrimaryButton(
+                  title: Text("Créer son instance soi-même",
                       style: TextStyle(
                           fontFamily: 'Roboto',
                           fontWeight: FontWeight.w500,
                           fontSize: 17)),
-                  onPressed: () => Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => Login()),
-                  ),
+                  onPressed: _launchURL,
+                  isEnabled: true,
                 ),
                 PrimaryButton(
-                  title: Text("Je n'ai pas d'instance",
+                  title: Text("Retour à la connexion",
                       style: TextStyle(
                           fontFamily: 'Roboto',
                           fontWeight: FontWeight.w500,
                           fontSize: 17)),
                   isEnabled: true,
-                  onPressed: () => Navigator.push(
-                    context,
-                    MaterialPageRoute(builder: (context) => NoInstance()),
-                  ),
+                  onPressed: () => Navigator.pop(context),
                 ),
               ],
             ),