+++ lib/ui/shared/sizeConfig.dart
@@ -1,26 +1,28 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/widgets.dart';
 
 class SizeConfig {
+  static MediaQueryData _mediaQueryData;
   static double screenWidth;
   static double screenHeight;
-  static double blockSizeHorizontal = 0;
-  static double blockSizeVertical = 0;
+  static double blockSizeHorizontal;
+  static double blockSizeVertical;
+  static double _safeAreaHorizontal;
+  static double _safeAreaVertical;
+  static double safeBlockHorizontal;
+  static double safeBlockVertical;
 
-  static double textMultiplier;
-  static double imageSizeMultiplier;
-  static double heightMultiplier;
-  static double widthMultiplier;
-
-  void init(BoxConstraints constraints) {
-    screenWidth = constraints.maxWidth;
-    screenHeight = constraints.maxHeight;
-
+  void init(BuildContext context) {
+    _mediaQueryData = MediaQuery.of(context);
+    screenWidth = _mediaQueryData.size.width;
+    screenHeight = _mediaQueryData.size.height;
     blockSizeHorizontal = screenWidth / 100;
     blockSizeVertical = screenHeight / 100;
-
-    textMultiplier = blockSizeVertical;
-    imageSizeMultiplier = blockSizeHorizontal;
-    heightMultiplier = blockSizeVertical;
-    widthMultiplier = blockSizeHorizontal;
+    _safeAreaHorizontal =
+        _mediaQueryData.padding.left + _mediaQueryData.padding.right;
+    _safeAreaVertical =
+        _mediaQueryData.padding.top + _mediaQueryData.padding.bottom;
+    safeBlockHorizontal = (screenWidth - _safeAreaHorizontal) / 100;
+    safeBlockVertical = (screenHeight - _safeAreaVertical) / 100;
   }
 }