+++ lib/ui/widgets/statcard.dart
@@ -1,0 +1,82 @@
+import 'package:flutter/foundation.dart';
+import 'package:flutter/material.dart';
+import 'package:sct_mobile/ui/shared/styles.dart';
+import 'package:sct_mobile/ui/widgets/cardlogo.dart';
+import 'package:sct_mobile/ui/shared/sizeConfig.dart';
+
+class StatCard extends StatefulWidget {
+  StatCard(
+      {@required this.title,
+      @required this.icon,
+      this.value,
+      @required this.color});
+  final String title;
+  final String icon;
+  final int value;
+  final int color;
+
+  @override
+  _StatCardState createState() => _StatCardState();
+}
+
+class _StatCardState extends State<StatCard> {
+  @override
+  Widget build(BuildContext context) {
+    SizeConfig().init(context);
+    return Card(
+      child: Row(
+        children: <Widget>[
+          SizedBox(height: 10.0, width: 10.0),
+          Container(
+              width: SizeConfig.blockSizeHorizontal * 70,
+              height: 150,
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: <Widget>[
+                  Expanded(
+                    child: Center(
+                      child: CardLogo(icon: widget.icon, color: widget.color),
+                    ),
+                  ),
+                  Expanded(
+                    child: Column(
+                      mainAxisAlignment: MainAxisAlignment.center,
+                      children: <Widget>[
+                        Text(
+                          widget.value.toString(),
+                          style: sctSubtitleStyleMedBold_b,
+                        ),
+                        Center(
+                          child: Text(
+                            widget.title,
+                            style: sctText,
+                            textAlign: TextAlign.center,
+                          ),
+                        ),
+                      ],
+                    ),
+                  ),
+                ],
+              )),
+        ],
+      ),
+      elevation: 5.0,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.only(
+          topRight: Radius.circular(25.0),
+          topLeft: Radius.circular(25.0),
+          bottomRight: Radius.circular(25.0),
+          bottomLeft: Radius.circular(25.0),
+        ),
+      ),
+    );
+  }
+}
+
+// SHADOW PROPERTIES
+// x0 y2 blur5
+// 30% black
+
+// SHADOW PROPER ACTIVE
+// x0 y1 blur5
+// 30% black